requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
timm>=0.9.0
transformers>=4.30.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
albumentations>=1.3.0
Pillow>=9.5.0
scikit-image>=0.21.0
imagehash>=4.3.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization and Plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tensorboard>=2.13.0
wandb>=0.15.0

# Interpretability and Explainability
shap>=0.42.0
lime>=0.2.0.1
captum>=0.6.0
grad-cam>=1.4.0

# Medical Image Processing
pydicom>=2.4.0
SimpleITK>=2.2.0
nibabel>=5.1.0

# Data Augmentation and Bias Mitigation
imgaug>=0.4.0
kornia>=0.7.0
fairlearn>=0.9.0

# Model Optimization and Deployment
onnx>=1.14.0
onnxruntime>=1.15.0
tensorrt>=8.6.0
openvino>=2023.0.0

# Utilities
tqdm>=4.65.0
rich>=13.4.0
hydra-core>=1.3.0
omegaconf>=2.3.0
python-dotenv>=1.0.0

# Testing and Quality
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Clinical Integration
streamlit>=1.25.0
gradio>=3.35.0
fastapi>=0.100.0
uvicorn>=0.22.0

# Database and Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.6.0
boto3>=1.28.0